Corrections on SBFem source code related to REAL/STATE typedef using.

+++ Projects/SBFem2D/mainsqrt.cpp
@@ -166,9 +166,9 @@ int main(int argc, char *argv[])
                 TPZManVector<double> eigval = celgrp->EigenvaluesReal();
                 TPZFMatrix<double> coef = celgrp->CoeficientsReal();
                 for (int i=0; i<eigval.size(); i++) {
-                    eigmap.insert(std::pair<double,double>(eigval[i],coef(i,0)));
+                    eigmap.insert(std::pair<REAL,REAL>(eigval[i],coef(i,0)));
                 }
-                for (std::multimap<double, double>::reverse_iterator it = eigmap.rbegin(); it!=eigmap.rend(); it++) {
+                for (std::multimap<REAL, REAL>::reverse_iterator it = eigmap.rbegin(); it!=eigmap.rend(); it++) {
                     results << it->first << "|" << it->second << " ";
                 }
             }
